Make gem sidekiq compliant.
Sidekiq 7 only accepts pure JSON as arguments. One of the issues is that
the i18n.locale calls returns a symbol, so, convert that symbol to a string
instead. Later, when in the base job, convert the locale back to a symbol for the
i81n.with_locale call.

Also, the notification name is returned as a symbol, which is later called
within a job, so again, falls foul of the Sidekiq 7 arguments issue.
We convert the symbol to a string, to pass through the job arguments.
Later, in the job the notification is sent to a send() call, which accepts
both strings and symbols, we just leave this a string never converting it back.